--- a/src/app/time-tracking/page.tsx
+++ b/src/app/time-tracking/page.tsx
@@ -1,0 +1,10 @@
+import { PageLayout } from "@/components/PageLayout"
+import { TimeTrackingDashboard } from "@/features/time-tracking/components/TimeTrackingDashboard"
+
+export default function TimeTrackingPage() {
+  return (
+    <PageLayout>
+      <TimeTrackingDashboard />
+    </PageLayout>
+  )
+}