… TIME_TRACKING_OPTIMIZATIONS.md
… package.json
… src/features/time-tracking/components/TimeTrackingDashboard.tsx
… src/features/time-tracking/hooks/use-time-tracking-optimized.ts